Microsoft Excel Updates
Statistics class can now parse through the WorkoutTracker class and record the max values for weight, reps, volume, and  the one rep max estimation onto an Excel document.

The document can currently display one graph (weight and one rep max) and I'm currently working on creating the volume and reps graph since the values are widely different.

Changed the Set constructor's parameter from int to double for its reps in order to coincide with the instance variable.

+++ FitnessAppJavaFX/src/model/Set.java
@@ -27,7 +27,7 @@ public class Set {
 	 * @param reps
 	 * @param weight
 	 */
-	public Set(int reps, double weight) {
+	public Set(double reps, double weight) {
 		this.reps = reps;
 		this.weight = weight;
 	}
